Fix dependency on `lines-and-columns`.
I messed up the minor updates I was making to `lines-and-columns` to use ES modules only, and published a patch update (v1.1.7) instead of a major one (v2.0.0). This had way more downstream effects than I intended. As I couldn't unpublish, I ended up publishing v1.1.9 to fix the issue, but that also included ESM support, but I realize now that was probably a mistake too since it represents a breaking change for anything using ESM itself.

This fixes it for `parse-json` by updating to the v2.0.2 ESM version.